+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/utils/extensions/String.kt
@@ -1,0 +1,19 @@
+package com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions
+
+fun String.splitByStep(step: Int): List<String> {
+    return (0 until length step step).map {
+        val currentLength = (it + step).let {
+            if (it > length) {
+                length
+            } else {
+                it
+            }
+        }
+        substring(it until currentLength)
+    }
+}
+
+private const val maxMessageSymbols = 4095
+
+fun String.splitForMessage(): List<String> = splitByStep(maxMessageSymbols)
+fun String.splitForMessageWithAdditionalStep(addStep: Int): List<String> = splitByStep(maxMessageSymbols + addStep)